fix : 파일구조 수정, favorite 이름 변경, WorkSpace, MindMap 추가 #15 (#43)
1. QuestionBoard, AnswerBoard domain 파일로 빼기
2. favorite -> bookmark로 이름 변경
3. WorkSpace, MindMap Entity 추가
