+++ src/main/java/com/solver/solver_be/domain/answerBoard/dto/AnswerRequestDto.java
@@ -1,0 +1,13 @@
+package com.solver.solver_be.domain.answerBoard.dto;
+
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class AnswerRequestDto {
+
+    private String contents;
+
+}
