--- src/main/java/com/solver/solver_be/domain/board/answerBoard/dto/AnswerRequestDto.java
@@ -1,13 +1,0 @@
-package com.solver.solver_be.domain.board.answerBoard.dto;
-
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class AnswerRequestDto {
-
-    private String contents;
-
-}
