--- src/main/java/com/solver/solver_be/domain/board/questionBoard/dto/KeywordRequestDto.java
@@ -1,10 +1,0 @@
-package com.solver.solver_be.domain.board.questionBoard.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class KeywordRequestDto {
-    String keyword;
-}
