+++ src/main/java/com/solver/solver_be/domain/mindmap/controller/MindMapController.java
@@ -1,0 +1,4 @@
+package com.solver.solver_be.domain.mindmap.controller;
+
+public class MindMapController {
+}
