+++ src/main/java/com/solver/solver_be/domain/mindmap/entity/MindMap.java
@@ -1,0 +1,4 @@
+package com.solver.solver_be.domain.mindmap.entity;
+
+public class MindMap {
+}
