+++ src/main/java/com/solver/solver_be/domain/mindmap/service/MindMapService.java
@@ -1,0 +1,4 @@
+package com.solver.solver_be.domain.mindmap.service;
+
+public class MindMapService {
+}
