+++ src/main/java/com/solver/solver_be/domain/questionBoard/dto/KeywordRequestDto.java
@@ -1,0 +1,10 @@
+package com.solver.solver_be.domain.questionBoard.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class KeywordRequestDto {
+    String keyword;
+}
