+++ src/main/java/com/solver/solver_be/domain/workspace/controller/WorkSpaceController.java
@@ -1,0 +1,4 @@
+package com.solver.solver_be.domain.workspace.controller;
+
+public class WorkSpaceController {
+}
