+++ src/main/java/com/solver/solver_be/domain/workspace/entity/WorkSpace.java
@@ -1,0 +1,4 @@
+package com.solver.solver_be.domain.workspace.entity;
+
+public class WorkSpace {
+}
