+++ src/main/java/com/solver/solver_be/domain/workspace/service/WorkSpaceService.java
@@ -1,0 +1,4 @@
+package com.solver.solver_be.domain.workspace.service;
+
+public class WorkSpaceService {
+}
